requirements.txt
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pandas>=2.0.0
streamlit>=1.35.0
